+++ CSharp/DevVmPowershell/Helpers/Implementations/ApplicationInstallHelper.cs
@@ -15,30 +15,24 @@ namespace Helpers.Implementations
 	public class ApplicationInstallHelper : IApplicationInstallHelper
 	{
 		private ServiceFactory ServiceFactory { get; }
-
+		private IConnectionHelper ConnectionHelper { get; }
 		private IWorkspaceHelper WorkspaceHelper { get; }
 		private IRestHelper RestHelper { get; }
 		private IRetryLogicHelper RetryLogicHelper { get; }
 
-		private string InstanceAddress { get; }
-		private string AdminUsername { get; }
-		private string AdminPassword { get; }
-
-		public ApplicationInstallHelper(IConnectionHelper connectionHelper, IRestHelper restHelper, IWorkspaceHelper workspaceHelper, IRetryLogicHelper retryLogicHelper, string instanceAddress, string adminUsername, string adminPassword)
+		public ApplicationInstallHelper(IConnectionHelper connectionHelper, IRestHelper restHelper, IWorkspaceHelper workspaceHelper, IRetryLogicHelper retryLogicHelper)
 		{
+			ConnectionHelper = connectionHelper;
 			ServiceFactory = connectionHelper.GetServiceFactory();
 			RestHelper = restHelper;
 			WorkspaceHelper = workspaceHelper;
 			RetryLogicHelper = retryLogicHelper;
-			InstanceAddress = instanceAddress;
-			AdminUsername = adminUsername;
-			AdminPassword = adminPassword;
 		}
 
 
 		public async Task<bool> InstallApplicationFromRapFileAsync(string workspaceName, string filePath)
 		{
-			HttpClient httpClient = RestHelper.GetHttpClient(InstanceAddress, AdminUsername, AdminPassword);
+			HttpClient httpClient = RestHelper.GetHttpClient(ConnectionHelper.RelativityInstanceName, ConnectionHelper.RelativityAdminUserName, ConnectionHelper.RelativityAdminPassword);
 
 			// Need to install to library before we install to a workspace
 			HttpResponseMessage updateResponseMessage = await UploadLibraryApplicationAsync(httpClient, filePath);
@@ -79,7 +73,7 @@ namespace Helpers.Implementations
 
 		public async Task<bool> InstallApplicationFromApplicationLibraryAsync(string workspaceName, string applicationGuid)
 		{
-			HttpClient httpClient = RestHelper.GetHttpClient(InstanceAddress, AdminUsername, AdminPassword);
+			HttpClient httpClient = RestHelper.GetHttpClient(ConnectionHelper.RelativityInstanceName, ConnectionHelper.RelativityAdminUserName, ConnectionHelper.RelativityAdminPassword);
 
 			int applicationId = GetApplicationLibraryIdAsync(applicationGuid).Result;
 
@@ -99,7 +93,7 @@ namespace Helpers.Implementations
 
 		private async Task<List<LibraryApplicationResponse>> ReadAllLibraryApplicationAsync()
 		{
-			HttpClient httpClient = RestHelper.GetHttpClient(InstanceAddress, AdminUsername, AdminPassword);
+			HttpClient httpClient = RestHelper.GetHttpClient(ConnectionHelper.RelativityInstanceName, ConnectionHelper.RelativityAdminUserName, ConnectionHelper.RelativityAdminPassword);
 			HttpResponseMessage httpResponse = await RestHelper.MakeGetAsync(httpClient, Constants.Connection.RestUrlEndpoints.ApplicationInstall.readAllLibraryApplicationUrl);
 
 			string content = await httpResponse.Content.ReadAsStringAsync();