+++ CSharp/DevVmPowershell/Helpers/Implementations/DisclaimerAcceptanceHelper.cs
@@ -17,23 +17,17 @@ namespace Helpers.Implementations
 {
 	public class DisclaimerAcceptanceHelper : IDisclaimerAcceptanceHelper
 	{
-		private IRSAPIClient RsapiClient { get; }
+		private IConnectionHelper ConnectionHelper { get; }
 		private IObjectManager ObjectManager { get; }
 		private ServiceFactory ServiceFactory { get; }
-		private string InstanceAddress { get; set; }
-		private string AdminUsername { get; set; }
-		private string AdminPassword { get; set; }
 		private IRestHelper RestHelper { get; set; }
 		private IWorkspaceHelper WorkspaceHelper { get; set; }
 
-		public DisclaimerAcceptanceHelper(IConnectionHelper connectionHelper, IRestHelper restHelper, IWorkspaceHelper workspaceHelper, string instanceAddress, string adminUsername, string adminPassword)
+		public DisclaimerAcceptanceHelper(IConnectionHelper connectionHelper, IRestHelper restHelper, IWorkspaceHelper workspaceHelper)
 		{
+			ConnectionHelper = connectionHelper;
 			ServiceFactory = connectionHelper.GetServiceFactory();
-			RsapiClient = ServiceFactory.CreateProxy<IRSAPIClient>();
 			ObjectManager = ServiceFactory.CreateProxy<IObjectManager>();
-			InstanceAddress = instanceAddress;
-			AdminUsername = adminUsername;
-			AdminPassword = adminPassword;
 			RestHelper = restHelper;
 			WorkspaceHelper = workspaceHelper;
 		}
@@ -56,7 +50,7 @@ namespace Helpers.Implementations
 		public async Task<bool> CheckIfDisclaimerConfigurationRdoExistsAsync(int workspaceId)
 		{
 			string url = Constants.Connection.RestUrlEndpoints.ObjectManager.QuerySlimUrl.Replace("-1", workspaceId.ToString());
-			HttpClient httpClient = RestHelper.GetHttpClient(InstanceAddress, AdminUsername, AdminPassword);
+			HttpClient httpClient = RestHelper.GetHttpClient(ConnectionHelper.RelativityInstanceName, ConnectionHelper.RelativityAdminUserName, ConnectionHelper.RelativityAdminPassword);
 			int objectTypeId = await GetDisclaimerSolutionConfigurationObjectTypeIdAsync(workspaceId);
 			var queryPayloadObject = new
 			{
@@ -103,7 +97,7 @@ namespace Helpers.Implementations
 		public async Task<bool> CheckIfDisclaimerRdoExistsAsync(int workspaceId)
 		{
 			string url = Constants.Connection.RestUrlEndpoints.ObjectManager.QuerySlimUrl.Replace("-1",workspaceId.ToString());
-			HttpClient httpClient = RestHelper.GetHttpClient(InstanceAddress, AdminUsername, AdminPassword);
+			HttpClient httpClient = RestHelper.GetHttpClient(ConnectionHelper.RelativityInstanceName, ConnectionHelper.RelativityAdminUserName, ConnectionHelper.RelativityAdminPassword);
 			int objectTypeId = await GetDisclaimerObjectTypeIdAsync(workspaceId);
 			var queryPayloadObject = new
 			{
@@ -134,7 +128,7 @@ namespace Helpers.Implementations
 
 		private async Task CreateDisclaimerConfigurationRdoAsync(int objectTypeId, int layoutId, int workspaceId)
 		{
-			HttpClient httpClient = RestHelper.GetHttpClient(InstanceAddress, AdminUsername, AdminPassword);
+			HttpClient httpClient = RestHelper.GetHttpClient(ConnectionHelper.RelativityInstanceName, ConnectionHelper.RelativityAdminUserName, ConnectionHelper.RelativityAdminPassword);
 			string url = $"Relativity.REST/api/Relativity.Objects/workspace/{workspaceId}/object/create";
 			ObjectManagerCreateRequestModel objectManagerCreateRequestModel = new ObjectManagerCreateRequestModel
 			{
@@ -193,7 +187,7 @@ namespace Helpers.Implementations
 
 		private async Task CreateDisclaimerRdoAsync(int objectTypeId, int layoutId, int workspaceId)
 		{
-			HttpClient httpClient = RestHelper.GetHttpClient(InstanceAddress, AdminUsername, AdminPassword);
+			HttpClient httpClient = RestHelper.GetHttpClient(ConnectionHelper.RelativityInstanceName, ConnectionHelper.RelativityAdminUserName, ConnectionHelper.RelativityAdminPassword);
 			string url = $"Relativity.REST/api/Relativity.Objects/workspace/{workspaceId}/object/create";
 			ObjectManagerCreateRequestModel objectManagerCreateRequestModel = new ObjectManagerCreateRequestModel
 			{
@@ -272,7 +266,7 @@ namespace Helpers.Implementations
 			try
 			{
 				string url = Constants.Connection.RestUrlEndpoints.ObjectManager.QuerySlimUrl.Replace("-1", workspaceId.ToString());
-				HttpClient httpClient = RestHelper.GetHttpClient(InstanceAddress, AdminUsername, AdminPassword);
+				HttpClient httpClient = RestHelper.GetHttpClient(ConnectionHelper.RelativityInstanceName, ConnectionHelper.RelativityAdminUserName, ConnectionHelper.RelativityAdminPassword);
 				var queryPayloadObject = new
 				{
 					request = new
@@ -314,7 +308,7 @@ namespace Helpers.Implementations
 			try
 			{
 				string url = Constants.Connection.RestUrlEndpoints.ObjectManager.QuerySlimUrl.Replace("-1", workspaceId.ToString());
-				HttpClient httpClient = RestHelper.GetHttpClient(InstanceAddress, AdminUsername, AdminPassword);
+				HttpClient httpClient = RestHelper.GetHttpClient(ConnectionHelper.RelativityInstanceName, ConnectionHelper.RelativityAdminUserName, ConnectionHelper.RelativityAdminPassword);
 				var queryPayloadObject = new
 				{
 					request = new
@@ -357,7 +351,7 @@ namespace Helpers.Implementations
 			try
 			{
 				string url = Constants.Connection.RestUrlEndpoints.ObjectManager.QuerySlimUrl.Replace("-1", workspaceId.ToString());
-				HttpClient httpClient = RestHelper.GetHttpClient(InstanceAddress, AdminUsername, AdminPassword);
+				HttpClient httpClient = RestHelper.GetHttpClient(ConnectionHelper.RelativityInstanceName, ConnectionHelper.RelativityAdminUserName, ConnectionHelper.RelativityAdminPassword);
 				var queryPayloadObject = new
 				{
 					request = new
@@ -398,7 +392,7 @@ namespace Helpers.Implementations
 			try
 			{
 				string url = Constants.Connection.RestUrlEndpoints.ObjectManager.QuerySlimUrl.Replace("-1", workspaceId.ToString());
-				HttpClient httpClient = RestHelper.GetHttpClient(InstanceAddress, AdminUsername, AdminPassword);
+				HttpClient httpClient = RestHelper.GetHttpClient(ConnectionHelper.RelativityInstanceName, ConnectionHelper.RelativityAdminUserName, ConnectionHelper.RelativityAdminPassword);
 				var queryPayloadObject = new
 				{
 					request = new